Fix policies for all routes & Organise helper methods
