+++ plugins/email_service/app/helpers/email_service/configset_helper.rb
@@ -1,0 +1,106 @@
+module EmailService
+  module ConfigsetHelper
+    include AwsSesHelper
+    include PlainEmailHelper
+    include TemplatedEmailHelper
+
+    def configset_create(name)
+      status = ""
+      begin
+        ses_client = create_ses_client
+        resp = ses_client.create_configuration_set({
+          configuration_set: { # required
+            name: name, # required
+          },
+        })
+        status = "success" 
+      rescue Aws::SES::Errors::ServiceError => error
+        status = "#{error}"
+        logger.debug "CRONUS: DEBUG: CONFIGSET: #{error}"
+      end
+
+      status
+    end
+
+    def get_configset
+      status = ""
+      configsets = []
+      begin
+        ses_client = create_ses_client
+        resp = ses_client.list_configuration_sets({
+          next_token: "",
+          max_items: 1000,
+        })
+
+        for index in 0 ... resp.configuration_sets.size
+          configsets << resp.configuration_sets[index].name
+        end if resp.configuration_sets.size > 0
+
+      rescue Aws::SES::Errors::ServiceError => error
+        status = "#{error}"
+        logger.debug "CRONUS: DEBUG: (AWS SES HELPER) CONFIGSET: #{error}"
+      end
+
+      # configsets if configsets && !configsets.empty?
+      configsets && !configsets.empty? ? configsets : error
+    # resp.configuration_sets #=> Array
+    # resp.configuration_sets[0].name #=> String
+    # resp.next_token #=> String
+    end
+
+    # Find existing config set
+    def find_configset(name)
+      resp = get_configset
+      resp.configuration_sets.each do |config_set|
+        logger.debug config_set[0].name
+      end
+    end
+
+    def describe_configset(name)
+      begin
+        ses_client = create_ses_client
+        resp = ses_client.describe_configuration_set({
+          configuration_set_name: name, # required
+          configuration_set_attribute_names: ["eventDestinations"], # accepts eventDestinations, trackingOptions, deliveryOptions, reputationOptions
+        })
+      rescue Aws::SES::Errors::ServiceError => error
+        status = "#{error}"
+        logger.debug "CRONUS: DEBUG: (AWS SES HELPER) DESCRIBE CONFIGSET: #{error}"
+      end
+      # resp.configuration_set.name #=> String
+      # resp.event_destinations #=> Array
+      # resp.event_destinations[0].name #=> String
+      # resp.event_destinations[0].enabled #=> Boolean
+      # resp.event_destinations[0].matching_event_types #=> Array
+      # resp.event_destinations[0].matching_event_types[0] #=> String, one of "send", "reject", "bounce", "complaint", "delivery", "open", "click", "renderingFailure"
+      # resp.event_destinations[0].kinesis_firehose_destination.iam_role_arn #=> String
+      # resp.event_destinations[0].kinesis_firehose_destination.delivery_stream_arn #=> String
+      # resp.event_destinations[0].cloud_watch_destination.dimension_configurations #=> Array
+      # resp.event_destinations[0].cloud_watch_destination.dimension_configurations[0].dimension_name #=> String
+      # resp.event_destinations[0].cloud_watch_destination.dimension_configurations[0].dimension_value_source #=> String, one of "messageTag", "emailHeader", "linkTag"
+      # resp.event_destinations[0].cloud_watch_destination.dimension_configurations[0].default_dimension_value #=> String
+      # resp.event_destinations[0].sns_destination.topic_arn #=> String
+      # resp.tracking_options.custom_redirect_domain #=> String
+      # resp.delivery_options.tls_policy #=> String, one of "Require", "Optional"
+      # resp.reputation_options.sending_enabled #=> Boolean
+      # resp.reputation_options.reputation_metrics_enabled #=> Boolean
+      # resp.reputation_options.last_fresh_start #=> Time
+    end
+
+    def configset_destroy(name)
+      status = ""
+      begin
+        ses_client = create_ses_client
+        resp = ses_client.delete_configuration_set({
+              configuration_set_name: name,
+        })
+        status = "success"
+      rescue Aws::SES::Errors::ServiceError => error
+        status = "#{error}"
+        logger.debug "CRONUS: DEBUG: (AWS SES HELPER) CONFIGSET: #{error}"
+      end
+      status
+    end
+
+  end
+end
