+++ plugins/email_service/app/views/email_service/verifications/_table_verified_email_item.html.haml
@@ -1,9 +1,9 @@
-- if not all_email[:identity].include?('@activation.email.global.cloud.sap')
+- if not email_item[:identity].include?('@activation.email.global.cloud.sap')
   %tr
     %td
-      = all_email[:identity]
+      = email_item[:identity]
     %td
-      = all_email[:status]
+      = email_item[:status]
     %td.snug
-      - if current_user.is_allowed?("email_service:verified_email_delete", {target: { all_email: all_email, project: @active_project, scoped_domain_name: @scoped_domain_name}}) or current_user.is_allowed?("email_service:verified_email_update", {})
-        = render partial: 'email_item_actions', locals: { all_email: all_email, show_view: false }
+      - if current_user.is_allowed?("email_service:verified_email_delete", {target: { email_item: email_item, project: @active_project, scoped_domain_name: @scoped_domain_name}}) or current_user.is_allowed?("email_service:verified_email_update", {})
+        = render partial: 'email_item_actions', locals: { email_item: email_item, show_view: false }
