+++ plugins/email_service/app/views/email_service/verifications/_table_verified_emails.html.haml
@@ -10,7 +10,7 @@
       %tr
         %td{ colspan: 3 } No verified eMail address available
     - else
-      -@all_emails.each do | all_email |
-        = render partial: 'table_verified_email_item', locals: { all_email: all_email }
+      -@all_emails.each do | email_item |
+        = render partial: 'table_verified_email_item', locals: { email_item: email_item }
 -#= paginate verified_email, :theme => 'twitter-bootstrap-3', :remote => true
