+++ plugins/email_service/config/policy.json
@@ -26,5 +26,8 @@
   "email_service:template_get": "rule:email_role_required",
   "email_service:template_create": "rule:email_role_required",
   "email_service:template_delete": "rule:email_role_required",
-  "email_service:template_update": "rule:email_role_required"
+  "email_service:template_update": "rule:email_role_required",
+  "email_service:email_stats": "rule:email_role_required",
+  "email_service:email_info": "rule:email_role_required",
+  "email_service:web_list": "rule:email_role_required"
 }
