+++ plugins/identity/app/services/service_layer/identity_services/aws/verified_email.rb
@@ -1,0 +1,21 @@
+# # frozen_String_literal: true
+
+# module ServiceLayer
+#   module IdentityServices
+#     module AWS
+#       # email api implementation
+#       module VerifiedEmail
+
+#         def verified_email_map
+#           @verified_email ||= class_map_proc(::EmailService::VerifiedEmail)
+#         end
+
+#         def new_verified_email(attributes = {})
+#           verified_email_map.call(attributes)
+#         end
+
+#       end
+
+#     end
+#   end
+# end
